pyproject.toml
[tool.poetry]
name = "disasmBot"
version = "0.1.0"
description = "A discord bot that dissasembles executables"
authors = ["Gabriel"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
capstone = "^5.0.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git"}
pefile = "^2021.9.3"
jishaku = "^2.5.0"
aiofiles = "^0.8.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
